+++ src/errors.rs
@@ -1,3 +1,4 @@
+use bytestring::ByteString;
 use ntex_amqp_codec::{protocol, AmqpCodecError, ProtocolIdError};
 
 #[derive(Debug, Display, Clone)]
@@ -28,4 +29,42 @@ pub enum SaslConnectError {
     Sasl(protocol::SaslCode),
     ExpectedOpenFrame,
     Disconnected,
+}
+
+#[derive(Debug, Display)]
+#[display(fmt = "Link error: {:?} {:?} ({:?})", err, description, info)]
+pub struct LinkError {
+    err: protocol::LinkError,
+    description: Option<ByteString>,
+    info: Option<protocol::Fields>,
+}
+
+impl LinkError {
+    pub fn force_detach() -> Self {
+        LinkError {
+            err: protocol::LinkError::DetachForced,
+            description: None,
+            info: None,
+        }
+    }
+
+    pub fn description<T: AsRef<str>>(mut self, text: T) -> Self {
+        self.description = Some(ByteString::from(text.as_ref()));
+        self
+    }
+
+    pub fn set_description(mut self, text: ByteString) -> Self {
+        self.description = Some(text);
+        self
+    }
+}
+
+impl Into<protocol::Error> for LinkError {
+    fn into(self) -> protocol::Error {
+        protocol::Error {
+            condition: self.err.into(),
+            description: self.description,
+            info: self.info,
+        }
+    }
 }